expires.conf: don't attempt to match dynamic languages
nginx only supports matching a single location-block before it stops execution, so the rule would have to go into the location-block that passes .php files over to FastCGI, and by that point it's too much of a hassle. modern browsers know not to cache .php files anyway, so let's simply remove the file extension.